backend/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
MarkupSafe==1.1.1
nltk==3.6.2
numpy==1.20.2
pandas==1.2.4
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
tqdm==4.60.0
urllib3==1.26.4
vaderSentiment==3.3.2
Werkzeug==0.16.1
